+++ src/main/java/io/supertokens/inmemorydb/Start.java
@@ -39,6 +39,9 @@ import io.supertokens.pluginInterface.emailverification.sqlStorage.EmailVerifica
 import io.supertokens.pluginInterface.exceptions.QuitProgramFromPluginException;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
+import io.supertokens.pluginInterface.jwt.JWTSigningKeyInfo;
+import io.supertokens.pluginInterface.jwt.exceptions.DuplicateKeyIdException;
+import io.supertokens.pluginInterface.jwt.sqlstorage.JWTRecipeSQLStorage;
 import io.supertokens.pluginInterface.session.SessionInfo;
 import io.supertokens.pluginInterface.session.sqlStorage.SessionSQLStorage;
 import io.supertokens.pluginInterface.sqlStorage.TransactionConnection;
@@ -50,9 +53,10 @@ import javax.annotation.Nullable;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.SQLTransactionRollbackException;
+import java.util.List;
 
 public class Start implements SessionSQLStorage, EmailPasswordSQLStorage, EmailVerificationSQLStorage,
-        ThirdPartySQLStorage {
+        ThirdPartySQLStorage, JWTRecipeSQLStorage {
 
     private static final Object appenderLock = new Object();
     private static final String APP_ID_KEY_NAME = "app_id";
@@ -793,6 +797,32 @@ public class Start implements SessionSQLStorage, EmailPasswordSQLStorage, EmailV
         try {
             return ThirdPartyQueries.getUsersCount(this);
         } catch (SQLException e) {
+            throw new StorageQueryException(e);
+        }
+    }
+
+    @Override
+    public List<JWTSigningKeyInfo> getJWTSigningKeys_Transaction(TransactionConnection con) throws StorageQueryException {
+        Connection sqlCon = (Connection) con.getConnection();
+        try {
+            return JWTSigningQueries.getJWTSigningKeys_Transaction(this, sqlCon);
+        } catch (SQLException e) {
+            throw new StorageQueryException(e);
+        }
+    }
+
+    @Override
+    public void setJWTSigningKey_Transaction(TransactionConnection con, JWTSigningKeyInfo info)
+            throws StorageQueryException, DuplicateKeyIdException {
+        Connection sqlCon = (Connection) con.getConnection();
+        try {
+            JWTSigningQueries.setJWTSigningKeyInfo_Transaction(this, sqlCon, info);
+        } catch (SQLException e) {
+
+            if (e.getMessage().equals("[SQLITE_CONSTRAINT]  Abort due to constraint violation (UNIQUE constraint failed: "+ Config.getConfig(this).getJWTSigningKeysTable() +".key_id)")) {
+                throw new DuplicateKeyIdException();
+            }
+
             throw new StorageQueryException(e);
         }
     }