+++ src/main/java/io/supertokens/storageLayer/StorageLayer.java
@@ -27,6 +27,7 @@ import io.supertokens.pluginInterface.Storage;
 import io.supertokens.pluginInterface.authRecipe.AuthRecipeStorage;
 import io.supertokens.pluginInterface.emailpassword.sqlStorage.EmailPasswordSQLStorage;
 import io.supertokens.pluginInterface.emailverification.sqlStorage.EmailVerificationSQLStorage;
+import io.supertokens.pluginInterface.jwt.JWTRecipeStorage;
 import io.supertokens.pluginInterface.session.SessionStorage;
 import io.supertokens.pluginInterface.thirdparty.sqlStorage.ThirdPartySQLStorage;
 
@@ -150,6 +151,14 @@ public class StorageLayer extends ResourceDistributor.SingletonResource {
             throw new UnsupportedOperationException("");
         }
         return (ThirdPartySQLStorage) getInstance(main).storage;
+    }
+
+    public static JWTRecipeStorage getJWTRecipeStorage(Main main) {
+        if (getInstance(main) == null) {
+            throw new QuitProgramException("please call init() before calling getStorageLayer");
+        }
+
+        return (JWTRecipeStorage) getInstance(main).storage;
     }
 
     public boolean isInMemDb() {