+++ src/main/java/io/supertokens/webserver/Webserver.java
@@ -31,6 +31,8 @@ import io.supertokens.webserver.api.emailverification.GenerateEmailVerificationT
 import io.supertokens.webserver.api.emailverification.RevokeAllTokensForUserAPI;
 import io.supertokens.webserver.api.emailverification.UnverifyEmailAPI;
 import io.supertokens.webserver.api.emailverification.VerifyEmailAPI;
+import io.supertokens.webserver.api.jwt.JWKSAPI;
+import io.supertokens.webserver.api.jwt.JWTSigningAPI;
 import io.supertokens.webserver.api.session.*;
 import io.supertokens.webserver.api.thirdparty.GetUsersByEmailAPI;
 import io.supertokens.webserver.api.thirdparty.SignInUpAPI;
@@ -171,6 +173,8 @@ public class Webserver extends ResourceDistributor.SingletonResource {
         addAPI(new UsersAPI(main));
         addAPI(new RevokeAllTokensForUserAPI(main));
         addAPI(new UnverifyEmailAPI(main));
+        addAPI(new JWTSigningAPI(main));
+        addAPI(new JWKSAPI(main));
 
         // deprecated APIs:
         addAPI(new RecipeRouter(main, new io.supertokens.webserver.api.emailpassword.UsersAPI(main),