+++ src/test/java/io/supertokens/test/emailverification/api/UnverifyEmailAPITest2_8.java
@@ -24,7 +24,7 @@ import io.supertokens.pluginInterface.STORAGE_TYPE;
 import io.supertokens.storageLayer.StorageLayer;
 import io.supertokens.test.TestingProcessManager;
 import io.supertokens.test.Utils;
-import io.supertokens.test.httpRequest.HttpRequest;
+import io.supertokens.test.httpRequest.HttpRequestForTesting;
 import io.supertokens.test.httpRequest.HttpResponseException;
 import org.junit.*;
 import org.junit.rules.TestRule;
@@ -95,7 +95,8 @@ public class UnverifyEmailAPITest2_8 {
     }
 
     private JsonObject unverifyEmail(Main main, JsonObject body) throws IOException, HttpResponseException {
-        return HttpRequest.sendJsonPOSTRequest(main, "", "http://localhost:3567/recipe/user/email/verify/remove", body,
+        return HttpRequestForTesting
+                .sendJsonPOSTRequest(main, "", "http://localhost:3567/recipe/user/email/verify/remove", body,
                 1000,
                 1000, null, Utils.getCdiVersion2_8ForTests(), RECIPE_ID.EMAIL_VERIFICATION.toString());
     }